feat: update device model to handle different data types for temperature, humidity, and gas ppm

+++ lib/domain/entities/air_data.dart
@@ -1,7 +1,7 @@
 class AirData {
   final double temperature;
   final double humidity;
-  final double gasPpm;
+  final int gasPpm;
 
   AirData({
     required this.temperature,
